requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
opencv-python
python-pptx